% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_code.R
\name{code_gen}
\alias{code_gen}
\alias{code_gen.double}
\alias{code_gen.integer}
\alias{code_gen.logical}
\alias{code_gen.character}
\alias{code_gen.factor}
\alias{code_gen.Date}
\alias{code_gen.POSIXct}
\alias{code_gen.difftime}
\alias{code_gen.list}
\alias{code_gen.tbl_df}
\alias{code_gen.data.frame}
\alias{code_gen.matrix}
\alias{code_gen.array}
\title{Generate Code}
\usage{
code_gen(object)

\method{code_gen}{double}(object)

\method{code_gen}{integer}(object)

\method{code_gen}{logical}(object)

\method{code_gen}{character}(object)

\method{code_gen}{factor}(object)

\method{code_gen}{Date}(object)

\method{code_gen}{POSIXct}(object)

\method{code_gen}{difftime}(object)

\method{code_gen}{list}(object)

\method{code_gen}{tbl_df}(object)

\method{code_gen}{data.frame}(object)

\method{code_gen}{matrix}(object)

\method{code_gen}{array}(object)
}
\arguments{
\item{object}{a vector, tibble, data.frame, list, matrix, or array}
}
\value{
a string of code to build the object
}
\description{
Generates a String of Code to build an object
}
\examples{
t1 <- data.frame(id = 1:6, sex = rep(c("M", "F"),3))
l1 <- as.list(rnorm(12))
code_gen(t1)
code_gen(l1)
}
