% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_code.R
\name{gen_code}
\alias{gen_code}
\alias{gen_code.double}
\alias{gen_code.integer}
\alias{gen_code.logical}
\alias{gen_code.character}
\alias{gen_code.factor}
\alias{gen_code.Date}
\alias{gen_code.POSIXct}
\alias{gen_code.difftime}
\alias{gen_code.list}
\alias{gen_code.tbl_df}
\alias{gen_code.data.frame}
\alias{gen_code.matrix}
\alias{gen_code.array}
\title{Generate Code}
\usage{
gen_code(object)

\method{gen_code}{double}(object)

\method{gen_code}{integer}(object)

\method{gen_code}{logical}(object)

\method{gen_code}{character}(object)

\method{gen_code}{factor}(object)

\method{gen_code}{Date}(object)

\method{gen_code}{POSIXct}(object)

\method{gen_code}{difftime}(object)

\method{gen_code}{list}(object)

\method{gen_code}{tbl_df}(object)

\method{gen_code}{data.frame}(object)

\method{gen_code}{matrix}(object)

\method{gen_code}{array}(object)
}
\arguments{
\item{object}{a vector, tibble, data.frame, list, matrix, or array}
}
\value{
a string of code to build the object
}
\description{
Generate Code
}
\examples{
t1 <- data.frame(id = 1:6, sex = rep(c("M", "F"),3))
l1 <- as.list(rnorm(12))
gen_code(t1)
gen_code(l1)
}
